Immovable mechanic

+++ js/suit3-engine/tile.js
@@ -1,7 +1,7 @@
 let id = 0;
 
 export class Tile {
-  constructor(row, col, color, type = Tile.type.DEFAULT) {
+  constructor(suit3, row, col, color, type = Tile.type.DEFAULT) {
     this.id = ++id;
     this.view = null;
     this.color = color;
@@ -9,6 +9,7 @@ export class Tile {
     this.row = row;
     this.col = col;
 
+    this.suit3 = suit3;
     this.fallPath = {x: [], y: []};
     this.fallDelay = 0;
   }
@@ -29,7 +30,17 @@ export class Tile {
 
   // after match, for destroy animation.
   kill(matched, cb) {
-    this.view.kill(matched, cb);
+    if (this.movable()) {
+      this.color = Math.floor(Math.random() * this.suit3.colors);
+      this.type = Tile.type.DEFAULT;
+
+      this.view.kill(matched, cb);
+    } else {
+      this.color -= 1;
+      this.type = this.color === 0 ? Tile.type.DEFAULT : this.type;
+
+      this.view.kick(matched, cb);
+    }
   }
 
   // after destroy, for fall animation.
@@ -39,7 +50,6 @@ export class Tile {
 
   // after destroy, for reset position and type. Sync.
   revive() {
-    // sync reset view here
     this.view.revive();
   }
 
@@ -52,6 +62,10 @@ export class Tile {
   afterShuffle(cb) {
     this.view.afterShuffle(cb); // update pos
   }
+
+  movable() {
+    return this.type !== Tile.type.IMMOVABLE;
+  }
 }
 
 Tile.type = {
@@ -61,7 +75,8 @@ Tile.type = {
   DIAGONAL: 3, // Spawn on matched L pattern. Destroy all in same diagonals
   ROOK: 4, // Spawn on matched T or + pattern. Destroy all in same row and col
   BOMB: 5, // Spawn on matched 5 tiles. Destroy all around, king like
-  DROP: 6, // Destroys itself on any bottom cell.
+  DROP: 6, // Destroys itself on any bottom cell
+  IMMOVABLE: 7, // Destroys if any neighbor explodes
 
   // Dynamic (not used as type for tile). No spawn, just destroy behaviour description
   QUEEN: 10, // on swap VERTICAL+HORIZONTAL. Destroy like queen
